Refocus desktop tests on journey flows
