+++ crates/defra-agent-desktop/src/app/tests/chat.rs
@@ -1,381 +1,6 @@
 use super::*;
 
 #[test]
-fn desktop_app_renders_chat_activity_with_live_session_data() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-
-    let principal_resp = runtime.block_on(core.node().execute(
-        r#"mutation {
-            add_AgentPrincipal(input: {
-                agent_did: "did:defra:amy"
-                display_name: "Amy"
-                default_behavior_id: "amy-default"
-                enabled: true
-            }) { agent_did }
-        }"#,
-    ));
-    assert!(!principal_resp.has_errors());
-
-    let created =
-        runtime.block_on(core.create_conversation("did:defra:amy", Some("amy-default")))?;
-    runtime.block_on(core.submit_request(
-        &created.session_id,
-        "did:defra:amy",
-        "hello operator",
-        None,
-    ))?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-
-    let texts = render_once(&mut app, &ctx);
-
-    assert_eq!(app.state.activity, Activity::Chat);
-    assert_eq!(
-        app.state.chat.shell.selected_agent_did.as_deref(),
-        Some("did:defra:amy")
-    );
-    assert_eq!(
-        app.state.chat.shell.selected_session_id.as_deref(),
-        Some(created.session_id.as_str())
-    );
-    assert!(!texts.iter().any(|text| text.contains("Operator Console")));
-    assert!(texts.iter().any(|text| text.contains("hello operator")));
-    assert!(texts.iter().any(|text| text.contains("Amy")));
-    Ok(())
-}
-
-#[test]
-fn desktop_app_renders_request_only_transcript_fallback() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-
-    runtime.block_on(insert_agent_principal(
-        &core,
-        "did:defra:amy",
-        "Amy",
-        "amy-default",
-    ))?;
-    let created =
-        runtime.block_on(core.create_conversation("did:defra:amy", Some("amy-default")))?;
-    runtime.block_on(core.submit_request(
-        &created.session_id,
-        "did:defra:amy",
-        "request only transcript row",
-        None,
-    ))?;
-
-    let mut driver = build_driver(
-        Arc::clone(&runtime),
-        core,
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    driver.app.state.activity = Activity::Chat;
-    let texts = driver.render();
-
-    assert_eq!(
-        driver.app.state.chat.shell.selected_session_id.as_deref(),
-        Some(created.session_id.as_str())
-    );
-    assert!(texts
-        .iter()
-        .any(|text| text.contains("request only transcript row")));
-    assert!(texts.iter().any(|text| text.contains("waiting for claim")));
-    assert!(!texts.iter().any(|text| text.contains("Transcript Empty")));
-    Ok(())
-}
-
-#[test]
-fn desktop_app_chat_header_retry_and_export_use_transcript_state() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-
-    let principal_resp = runtime.block_on(core.node().execute(
-        r#"mutation {
-            add_AgentPrincipal(input: {
-                agent_did: "did:defra:amy"
-                display_name: "Amy"
-                default_behavior_id: "amy-default"
-                enabled: true
-            }) { agent_did }
-        }"#,
-    ));
-    assert!(!principal_resp.has_errors());
-
-    let created =
-        runtime.block_on(core.create_conversation("did:defra:amy", Some("amy-default")))?;
-    let submitted = runtime.block_on(core.submit_request(
-        &created.session_id,
-        "did:defra:amy",
-        "hello operator",
-        None,
-    ))?;
-    let response = runtime.block_on(core.node().execute(&format!(
-        r#"mutation {{
-            add_AgentResponse(input: {{
-                response_key: "response-retry-export-1"
-                request_id: "{request_id}"
-                agent_did: "did:defra:amy"
-                behavior_id: "amy-default"
-                session_id: "{session_id}"
-                content: "assistant complete"
-                reasoning: ""
-                status: "completed"
-                error_message: ""
-                token_count: 2
-                progress_seq: 1
-                created_at: "2026-04-14T00:00:04Z"
-                completed_at: "2026-04-14T00:00:05Z"
-            }}) {{ response_key }}
-        }}"#,
-        request_id = escape_graphql_string(&submitted.request_id),
-        session_id = escape_graphql_string(&created.session_id),
-    )));
-    assert!(!response.has_errors());
-    runtime.block_on(core.refresh_store())?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    app.state.activity = Activity::Chat;
-    let mut driver = AuditDriver::new(app, ctx);
-
-    let initial = driver.render();
-    assert!(initial.iter().any(|text| text.contains("Retry")));
-    assert!(initial.iter().any(|text| text.contains("Export")));
-    assert!(initial.iter().any(|text| text.contains("completed")));
-    assert!(driver.has_target(audit::targets::CHAT_RETRY));
-    assert!(driver.has_target(audit::targets::CHAT_EXPORT));
-
-    let selected_session_id = driver.app.state.chat.shell.selected_session_id.clone();
-    driver.click_interactable_target(audit::targets::CHAT_EXPORT)?;
-    let export_payload = driver
-        .app
-        .state
-        .chat
-        .editor
-        .last_export_payload
-        .as_deref()
-        .ok_or_else(|| anyhow!("chat export did not capture a payload"))?;
-    assert!(export_payload.contains("hello operator"));
-    assert!(export_payload.contains("assistant complete"));
-    assert!(export_payload.contains(&submitted.request_id));
-
-    driver.click_interactable_target(audit::targets::CHAT_RETRY)?;
-    let retry_request = driver
-        .app
-        .client
-        .as_ref()
-        .and_then(|client| {
-            client
-                .store()
-                .snapshot()
-                .requests
-                .iter()
-                .find(|row| row.retry_parent_request.as_deref() == Some(&submitted.request_id))
-                .cloned()
-        })
-        .ok_or_else(|| anyhow!("retry request was not created"))?;
-
-    assert_eq!(
-        driver.app.state.chat.shell.selected_session_id,
-        selected_session_id
-    );
-    assert_eq!(driver.app.state.chat.editor.last_submission_error, None);
-    assert_eq!(retry_request.content.as_deref(), Some("hello operator"));
-    assert_eq!(
-        retry_request.retry_root_request.as_deref(),
-        Some(submitted.request_id.as_str())
-    );
-    assert_eq!(retry_request.retry_count, Some(1));
-    Ok(())
-}
-
-#[test]
-fn desktop_app_renders_chat_first_conversation_nudge() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-
-    let principal_resp = runtime.block_on(core.node().execute(
-        r#"mutation {
-            add_AgentPrincipal(input: {
-                agent_did: "did:defra:amy"
-                display_name: "Amy"
-                default_behavior_id: "amy-default"
-                enabled: true
-            }) { agent_did }
-        }"#,
-    ));
-    assert!(!principal_resp.has_errors());
-    runtime.block_on(core.refresh_store())?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    app.state.activity = Activity::Chat;
-
-    let texts = render_once(&mut app, &ctx);
-
-    assert_eq!(app.state.activity, Activity::Chat);
-    assert_eq!(
-        app.state.chat.shell.selected_agent_did.as_deref(),
-        Some("did:defra:amy")
-    );
-    assert!(app.state.chat.shell.selected_session_id.is_some());
-    assert!(texts.iter().any(|text| text.contains("Transcript Empty")));
-    assert!(!texts
-        .iter()
-        .any(|text| text.contains("Automatic conversation creation did not complete")));
-    Ok(())
-}
-
-#[test]
-fn desktop_app_auto_creates_first_chat_conversation() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-
-    let principal_resp = runtime.block_on(core.node().execute(
-        r#"mutation {
-            add_AgentPrincipal(input: {
-                agent_did: "did:defra:amy"
-                display_name: "Amy"
-                default_behavior_id: "amy-default"
-                enabled: true
-            }) { agent_did }
-        }"#,
-    ));
-    assert!(!principal_resp.has_errors());
-    runtime.block_on(core.refresh_store())?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    app.state.activity = Activity::Chat;
-    let mut driver = AuditDriver::new(app, ctx);
-
-    let texts = driver.render();
-
-    assert!(driver.app.state.chat.shell.selected_session_id.is_some());
-    assert!(texts.iter().any(|text| text.contains("Transcript Empty")));
-    assert!(!texts
-        .iter()
-        .any(|text| text.contains("Automatic conversation creation did not complete")));
-    Ok(())
-}
-
-#[test]
-fn desktop_app_clicks_through_activity_bar_navigation() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-    let log_store = Arc::new(DesktopLogStore::new(64));
-    log_store.record_manual(
-        chrono::Utc::now(),
-        tracing::Level::INFO,
-        "defra_agent_desktop::replication",
-        "activity navigation marker",
-        [("marker", "activity".to_string())],
-    );
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        log_store,
-    );
-    app.state.onboarding.first_launch_redirect_done = true;
-    app.state.activity = Activity::Chat;
-    let mut driver = AuditDriver::new(app, ctx);
-
-    let chat_texts = driver.render();
-    assert!(chat_texts.iter().any(|text| text.contains("DEFRA DESKTOP")));
-    assert!(chat_texts.iter().any(|text| text.contains("CHAT")));
-    assert!(chat_texts.iter().any(|text| text.contains("conversations")));
-    assert!(chat_texts.iter().any(|text| text.contains("LOGS")));
-    assert!(chat_texts
-        .iter()
-        .any(|text| text.contains("Add Deployment")));
-
-    let logs_texts = driver.open_activity(Activity::Logs);
-    assert_eq!(driver.app.state.activity, Activity::Logs);
-    assert!(logs_texts.iter().any(|text| text.contains("Live Logs")));
-    assert!(logs_texts.iter().any(|text| text.contains("Log Controls")));
-
-    let operator_texts = driver.open_activity(Activity::Operator);
-    assert_eq!(driver.app.state.activity, Activity::Operator);
-    assert!(operator_texts
-        .iter()
-        .any(|text| text.contains("Operator Console")));
-    assert!(operator_texts.iter().any(|text| text.contains("OPERATOR")));
-
-    let peers_texts = driver.open_activity(Activity::Peers);
-    assert_eq!(driver.app.state.activity, Activity::Peers);
-    assert!(peers_texts
-        .iter()
-        .any(|text| text.contains("Add Your First Deployment")));
-    assert!(peers_texts.iter().any(|text| text.contains("PEERS")));
-
-    let back_to_chat = driver.open_activity(Activity::Chat);
-    assert_eq!(driver.app.state.activity, Activity::Chat);
-    assert!(back_to_chat
-        .iter()
-        .any(|text| text.contains("Add Deployment")));
-    Ok(())
-}
-
-#[test]
 fn desktop_app_clicks_through_chat_deployment_and_conversation_switching() -> Result<()> {
     let runtime = test_runtime()?;
     let tempdir = tempfile::tempdir()?;
@@ -592,117 +217,6 @@ fn desktop_app_clicks_through_chat_reasoning_and_tool_card_disclosures() -> Resu
 }
 
 #[test]
-fn desktop_app_clicks_through_chat_send_without_precreating_conversation() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path().join("desktop")),
-        ClientCoreOptions::local_only(),
-    ))?;
-    let mock_endpoint = MockModelEndpoint::start("default")?;
-    let running_agent = runtime.block_on(spawn_backed_agent(
-        core.node_arc(),
-        tempdir.path().join("agent").join("audit-direct-send.key"),
-        "audit-direct-send",
-        &AgentBackendConfig::mock(mock_endpoint.endpoint()),
-    ))?;
-    runtime.block_on(core.refresh_store())?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    app.state.activity = Activity::Chat;
-    let mut driver = AuditDriver::new(app, ctx);
-
-    let _session_id = ensure_chat_session_selected(
-        &mut driver,
-        "chat session ready before direct send",
-        Duration::from_secs(5),
-    )?;
-    assert_eq!(
-        driver.app.state.chat.shell.selected_agent_did.as_deref(),
-        Some(running_agent.did.as_str())
-    );
-
-    driver.click_target(audit::targets::CHAT_COMPOSER_TEXT);
-    driver.type_text("send directly without creating the session first");
-    driver.render();
-    driver.click_target(audit::targets::CHAT_SEND);
-    if wait_for_value(
-        "session created by first direct-send click",
-        Duration::from_secs(1),
-        || driver.app.state.chat.shell.selected_session_id.clone(),
-    )
-    .is_err()
-    {
-        driver.render();
-        driver.click_target(audit::targets::CHAT_SEND);
-    }
-
-    let session_id = wait_for_value(
-        "session created by direct send",
-        Duration::from_secs(5),
-        || driver.app.state.chat.shell.selected_session_id.clone(),
-    )?;
-    assert!(driver.app.state.chat.editor.last_submission_error.is_none());
-    assert!(driver.app.state.chat.editor.composer_text.is_empty());
-
-    let request_id = wait_for_value(
-        "direct-send focused request id",
-        Duration::from_secs(5),
-        || {
-            driver
-                .app
-                .client
-                .as_ref()
-                .and_then(|client| client.store().focused_request_id())
-        },
-    )?;
-    wait_for_value(
-        "direct-send response row in store",
-        Duration::from_secs(10),
-        || {
-            driver.app.client.as_ref().and_then(|client| {
-                client
-                    .store()
-                    .snapshot()
-                    .latest_response_for_request(&request_id)
-                    .and_then(|row| row.content.as_deref())
-                    .filter(|content| !content.trim().is_empty())
-                    .map(str::to_string)
-            })
-        },
-    )?;
-    let transcript_texts = wait_for_value(
-        "direct-send transcript response",
-        Duration::from_secs(10),
-        || {
-            let texts = driver.render();
-            texts
-                .iter()
-                .any(|text| text.contains("mock response"))
-                .then_some(texts)
-        },
-    )?;
-    assert_eq!(
-        driver.app.state.chat.shell.selected_session_id.as_deref(),
-        Some(session_id.as_str())
-    );
-    assert!(transcript_texts
-        .iter()
-        .any(|text| { text.contains("send directly without creating the session first") }));
-
-    runtime.block_on(running_agent.shutdown())?;
-    Ok(())
-}
-
-#[test]
 fn desktop_app_blocks_chat_send_while_turn_is_waiting_for_claim() -> Result<()> {
     let runtime = test_runtime()?;
     let tempdir = tempfile::tempdir()?;
@@ -782,138 +296,6 @@ fn desktop_app_blocks_chat_send_while_turn_is_waiting_for_claim() -> Result<()> 
         "blocked follow-up"
     );
     assert_eq!(driver.app.state.chat.editor.last_submission_error, None);
-    Ok(())
-}
-
-#[test]
-fn desktop_app_clicks_through_live_agent_submission() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path().join("desktop")),
-        ClientCoreOptions::local_only(),
-    ))?;
-    let mock_endpoint = MockModelEndpoint::start("default")?;
-    let running_agent = runtime.block_on(spawn_backed_agent(
-        core.node_arc(),
-        tempdir.path().join("agent").join("audit-live.key"),
-        "audit-live",
-        &AgentBackendConfig::mock(mock_endpoint.endpoint()),
-    ))?;
-    runtime.block_on(core.refresh_store())?;
-
-    let ctx = egui::Context::default();
-    let cc = eframe::CreationContext::_new_kittest(ctx.clone());
-    let mut app = DesktopApp::from_parts(
-        &cc,
-        Arc::clone(&runtime),
-        Some(Arc::new(core)),
-        Vec::new(),
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    app.state.activity = Activity::Chat;
-    let mut driver = AuditDriver::new(app, ctx);
-
-    assert!(driver
-        .app
-        .client
-        .as_ref()
-        .expect("desktop client should exist")
-        .store()
-        .snapshot()
-        .agent_principals
-        .iter()
-        .any(|row| row.agent_did == running_agent.did));
-
-    let initial = driver.render();
-    ensure_chat_agent_selected(
-        &mut driver,
-        "chat agent selected for live submission",
-        Duration::from_secs(5),
-    )?;
-    assert_eq!(
-        driver.app.state.chat.shell.selected_agent_did.as_deref(),
-        Some(running_agent.did.as_str())
-    );
-    let _session_id = ensure_chat_session_selected(
-        &mut driver,
-        "chat session selected for live submission",
-        Duration::from_secs(5),
-    )?;
-    let after_create = driver.render();
-    assert!(driver.app.state.chat.shell.selected_session_id.is_some());
-    assert!(!initial
-        .iter()
-        .any(|text| text.contains("Automatic conversation creation did not complete")));
-    assert!(after_create
-        .iter()
-        .any(|text| text.contains("Transcript Empty")));
-
-    driver.click_target(audit::targets::CHAT_COMPOSER_TEXT);
-    driver.type_text("say hello from the desktop audit");
-    assert_eq!(
-        driver.app.state.chat.editor.composer_text,
-        "say hello from the desktop audit"
-    );
-    driver.press_key(
-        egui::Key::Enter,
-        egui::Modifiers {
-            command: true,
-            mac_cmd: true,
-            ..Default::default()
-        },
-    );
-    assert_eq!(driver.app.state.chat.editor.last_submission_error, None);
-    assert!(driver.app.state.chat.editor.composer_text.is_empty());
-
-    let request_id = wait_for_value("focused request id", Duration::from_secs(5), || {
-        driver
-            .app
-            .client
-            .as_ref()
-            .and_then(|client| client.store().focused_request_id())
-    })?;
-    wait_for_value(
-        "response row in client store",
-        Duration::from_secs(10),
-        || {
-            driver.app.client.as_ref().and_then(|client| {
-                client
-                    .store()
-                    .snapshot()
-                    .latest_response_for_request(&request_id)
-                    .and_then(|row| row.content.as_deref())
-                    .filter(|content| !content.trim().is_empty())
-                    .map(str::to_string)
-            })
-        },
-    )?;
-    let response_texts = wait_for_value(
-        "mock response in transcript",
-        Duration::from_secs(10),
-        || {
-            let texts = driver.render();
-            texts
-                .iter()
-                .any(|text| text.contains("mock response"))
-                .then_some(texts)
-        },
-    )?;
-    assert!(response_texts
-        .iter()
-        .any(|text| text.contains("say hello from the desktop audit")));
-    assert!(driver
-        .app
-        .client
-        .as_ref()
-        .expect("desktop client should exist")
-        .store()
-        .snapshot()
-        .responses
-        .iter()
-        .any(|row| row.request_id.as_deref() == Some(request_id.as_str())));
-
-    runtime.block_on(running_agent.shutdown())?;
     Ok(())
 }
 
@@ -1163,46 +545,5 @@ fn desktop_app_clicks_through_live_agent_multi_turn_conversation() -> Result<()>
     assert!(final_texts.iter().any(|text| text.contains("completed")));
 
     runtime.block_on(running_agent.shutdown())?;
-    Ok(())
-}
-
-#[test]
-fn desktop_app_places_tool_cards_in_tool_turn_without_raw_duplicate_message() -> Result<()> {
-    let runtime = test_runtime()?;
-    let tempdir = tempfile::tempdir()?;
-    let core = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path()),
-        ClientCoreOptions::local_only(),
-    ))?;
-    runtime.block_on(insert_agent_principal(
-        &core,
-        "did:defra:amy",
-        "Amy",
-        "amy-default",
-    ))?;
-    let conversation =
-        runtime.block_on(core.create_conversation("did:defra:amy", Some("amy-default")))?;
-    runtime.block_on(insert_chat_transcript_documents(
-        &core,
-        &conversation.session_id,
-        "did:defra:amy",
-        "amy-default",
-        "response-tool-structure-1",
-    ))?;
-
-    let mut driver = build_driver(
-        Arc::clone(&runtime),
-        core,
-        Arc::new(DesktopLogStore::new(64)),
-    );
-    driver.app.state.activity = Activity::Chat;
-    let texts = driver.render();
-
-    assert!(texts.iter().any(|text| text.contains("TOOL")));
-    assert!(texts.iter().any(|text| text.contains("shell  completed")));
-    assert!(!texts
-        .iter()
-        .any(|text| text.contains("src/app.rs: audit target live")));
-    assert!(!texts.iter().any(|text| text.contains("\"call_id\"")));
     Ok(())
 }
