+++ crates/defra-agent-desktop/src/app/tests/live/chat.rs
@@ -14,16 +14,6 @@ fn desktop_app_live_inference_smoke() -> Result<()> {
         DesktopPaths::from_root(tempdir.path().join("desktop")),
         ClientCoreOptions::local_only(),
     ))?;
-    let peer = runtime.block_on(ClientCore::start_with_paths_and_options(
-        DesktopPaths::from_root(tempdir.path().join("peer")),
-        ClientCoreOptions::local_only(),
-    ))?;
-    let peer_addr = peer
-        .listen_addresses()
-        .first()
-        .cloned()
-        .ok_or_else(|| anyhow!("live peer missing listen address"))?;
-    let baseline_events = global_log_store().snapshot().total_events;
     let running_agent = runtime.block_on(spawn_backed_agent(
         core.node_arc(),
         tempdir.path().join("agent").join("audit-live.key"),
@@ -272,76 +262,6 @@ fn desktop_app_live_inference_smoke() -> Result<()> {
         &response_content,
         &first_tokens,
     )?;
-    let (
-        request_lifecycle_state,
-        response_status,
-        runtime_process_state,
-        runtime_default_behavior_id,
-        runtime_last_result,
-        runtime_runnable_behaviors,
-        runtime_scheduled_task_count,
-        _initial_live_store_row_count,
-    ) = wait_for_value(
-        "live operator rows available",
-        Duration::from_secs(10),
-        || {
-            driver.app.client.as_ref().and_then(|client| {
-                let snapshot = client.store().snapshot();
-                let request = snapshot
-                    .requests
-                    .iter()
-                    .find(|row| row.request_id == request_id)?;
-                let response = snapshot.latest_response_for_request(&request_id)?;
-                let runtime_row = snapshot.latest_runtime(&running_agent.did)?;
-                let backend_row = snapshot
-                    .inference_backends
-                    .iter()
-                    .find(|row| row.backend_id == live_backend_id)?;
-
-                Some((
-                    request
-                        .lifecycle_state
-                        .clone()
-                        .unwrap_or_else(|| "unset".to_string()),
-                    response
-                        .status
-                        .clone()
-                        .unwrap_or_else(|| "unset".to_string()),
-                    runtime_row
-                        .process_state
-                        .clone()
-                        .unwrap_or_else(|| "unknown".to_string()),
-                    runtime_row
-                        .default_behavior_id
-                        .clone()
-                        .unwrap_or_else(|| "unbound".to_string()),
-                    runtime_row
-                        .last_reconcile_result
-                        .clone()
-                        .unwrap_or_else(|| "pending".to_string()),
-                    runtime_row
-                        .runnable_behavior_count
-                        .unwrap_or_default()
-                        .to_string(),
-                    snapshot
-                        .scheduled_tasks
-                        .iter()
-                        .filter(|row| row.agent_did.as_deref() == Some(running_agent.did.as_str()))
-                        .count()
-                        .to_string(),
-                    snapshot.row_count().to_string(),
-                ))
-                .filter(|_| {
-                    backend_row.provider_kind.as_deref() == Some(backend.provider_kind.as_str())
-                        && backend_row.endpoint.as_deref() == Some(backend.endpoint.as_str())
-                        && backend_row
-                            .models
-                            .iter()
-                            .any(|model| model == &backend.model_name)
-                })
-            })
-        },
-    )?;
 
     let chat_texts = wait_for_value(
         "live response in first transcript",
@@ -428,224 +348,6 @@ fn desktop_app_live_inference_smoke() -> Result<()> {
         },
     )?;
 
-    driver.open_activity(Activity::Operator);
-    driver.click_target(&audit::targets::operator_section(
-        crate::state::OperatorSection::Runtime,
-    ));
-    let runtime_texts = wait_for_value(
-        "operator runtime inspector",
-        Duration::from_secs(10),
-        || {
-            let texts = driver.render();
-            texts
-                .iter()
-                .any(|text| text.contains("Runtime Inspector"))
-                .then_some(texts)
-        },
-    )?;
-    assert!(runtime_texts
-        .iter()
-        .any(|text| text.contains(&runtime_process_state)));
-    assert!(runtime_texts
-        .iter()
-        .any(|text| text.contains(&runtime_default_behavior_id)));
-    assert!(runtime_texts
-        .iter()
-        .any(|text| text.contains(&runtime_last_result)));
-    assert!(runtime_texts
-        .iter()
-        .any(|text| text.contains(&runtime_runnable_behaviors)));
-    assert!(runtime_texts
-        .iter()
-        .any(|text| text.contains(&runtime_scheduled_task_count)));
-
-    driver.click_target(&audit::targets::operator_section(
-        crate::state::OperatorSection::RequestTimeline,
-    ));
-    let operator_texts = wait_for_value(
-        "live request row in operator timeline",
-        Duration::from_secs(10),
-        || {
-            driver
-                .wait_for_target(
-                    "operator request row",
-                    Duration::from_millis(250),
-                    &audit::targets::operator_entity(&request_id),
-                )
-                .ok()?;
-            let texts = driver.click_target(&audit::targets::operator_entity(&request_id));
-            texts
-                .iter()
-                .any(|text| text.contains("Request Detail"))
-                .then_some(texts)
-        },
-    )?;
-    assert_eq!(
-        driver.app.state.operator.selected_entity_id.as_deref(),
-        Some(request_id.as_str())
-    );
-    assert!(operator_texts
-        .iter()
-        .any(|text| text.contains(&request_lifecycle_state)));
-    assert!(operator_texts
-        .iter()
-        .any(|text| text.contains(&response_status)));
-
-    driver.click_target(&audit::targets::operator_section(
-        crate::state::OperatorSection::Backends,
-    ));
-    driver.wait_for_target(
-        "live backend entity",
-        Duration::from_secs(10),
-        &audit::targets::operator_entity(&live_backend_id),
-    )?;
-    let backend_texts = driver.click_target(&audit::targets::operator_entity(&live_backend_id));
-    assert_eq!(
-        driver.app.state.operator.selected_entity_id.as_deref(),
-        Some(live_backend_id.as_str())
-    );
-    assert!(backend_texts
-        .iter()
-        .any(|text| text.contains("Provider Kind")));
-    assert!(backend_texts
-        .iter()
-        .any(|text| text.contains(backend.endpoint.as_str())));
-    match driver.app.state.operator.draft.as_ref() {
-        Some(OperatorDraft::Backend(draft)) => {
-            assert_eq!(draft.backend_id, live_backend_id);
-            assert_eq!(draft.provider_kind, backend.provider_kind.as_str());
-            assert_eq!(draft.endpoint, backend.endpoint);
-            assert_eq!(draft.models, backend.model_name);
-            assert!(draft.enabled);
-            assert_eq!(draft.max_queue_depth, "100");
-        }
-        other => panic!("expected backend draft in live smoke, got {other:?}"),
-    }
-
-    driver.open_activity(Activity::Peers);
-    let peers_texts = driver.wait_for_target(
-        "peers add-deployment form",
-        Duration::from_secs(10),
-        audit::targets::PEERS_ADD_LABEL,
-    )?;
-    assert!(peers_texts
-        .iter()
-        .any(|text| text.contains("Add Your First Deployment")));
-    driver.click_target(audit::targets::PEERS_ONBOARDING_COPY_DID);
-    assert_eq!(
-        driver.app.state.peers.last_action_message.as_deref(),
-        Some("Copied desktop DID to clipboard.")
-    );
-    driver.click_target(audit::targets::PEERS_ADD_LABEL);
-    driver.type_text("Scratch Remote");
-    driver.click_target(audit::targets::PEERS_ADD_ADDR);
-    driver.type_text("iroh://scratch-address");
-    driver.click_target(audit::targets::PEERS_ADD_AGENT_DID);
-    driver.type_text("did:defra:scratch");
-    driver.click_target(audit::targets::PEERS_CLEAR);
-    assert!(driver.app.state.peers.add_label.is_empty());
-    assert!(driver.app.state.peers.add_addr.is_empty());
-    assert!(driver.app.state.peers.add_agent_did.is_empty());
-    assert_eq!(driver.app.state.peers.last_action_message, None);
-    driver.click_target(audit::targets::PEERS_ADD_LABEL);
-    driver.type_text("Live Remote");
-    driver.click_target(audit::targets::PEERS_ADD_ADDR);
-    driver.type_text(&peer_addr);
-    driver.click_target(audit::targets::PEERS_ADD_AGENT_DID);
-    driver.type_text("did:defra:peer-live");
-    driver.click_target(audit::targets::PEERS_SAVE);
-    let live_peer_id = wait_for_value("live peer added", Duration::from_secs(10), || {
-        driver.app.client.as_ref().and_then(|client| {
-            let records = driver.app.runtime.block_on(client.peer_records());
-            records
-                .iter()
-                .find(|record| record.label == "Live Remote")
-                .map(|record| record.peer_id.clone())
-        })
-    })?;
-    let live_peer_target = audit::targets::peers_peer(&live_peer_id);
-    driver.wait_for_target(
-        "live peer row after save",
-        Duration::from_secs(10),
-        &live_peer_target,
-    )?;
-    driver.click_interactable_target(audit::targets::PEERS_MAIN_COPY_DID)?;
-    assert_eq!(
-        driver.app.state.peers.last_action_message.as_deref(),
-        Some("Copied desktop DID to clipboard.")
-    );
-    driver.click_target(&live_peer_target);
-    assert_eq!(
-        driver.app.state.peers.selected_peer_id.as_deref(),
-        Some(live_peer_id.as_str())
-    );
-    driver
-        .scroll_right_rail_until_target("peers remove saved peer", audit::targets::PEERS_REMOVE)?;
-    driver.click_interactable_target(audit::targets::PEERS_REMOVE)?;
-    wait_for_value("live peer removed", Duration::from_secs(10), || {
-        driver
-            .app
-            .client
-            .as_ref()
-            .filter(|client| client.configured_peer_count() == 0)
-            .map(|_| ())
-    })?;
-
-    global_log_store().record_manual(
-        chrono::Utc::now(),
-        tracing::Level::INFO,
-        "defra_agent_desktop::replication",
-        "live audit replication marker",
-        [("request_id", request_id.clone())],
-    );
-    global_log_store().record_manual(
-        chrono::Utc::now(),
-        tracing::Level::WARN,
-        "defra_agent_desktop::peer",
-        "live audit warning marker",
-        [("peer_id", "peer-live".to_string())],
-    );
-
-    let live_store_row_count = driver
-        .app
-        .client
-        .as_ref()
-        .map(|client| client.store().snapshot().row_count().to_string())
-        .ok_or_else(|| anyhow!("desktop client missing"))?;
-    let logs_texts = driver.open_activity(Activity::Logs);
-    assert!(logs_texts.iter().any(|text| text.contains("Live Logs")));
-    assert!(logs_texts.iter().any(|text| text.contains("approx store")));
-    assert!(logs_texts
-        .iter()
-        .any(|text| text.contains(&format!("/ {live_store_row_count} rows"))));
-    assert!(logs_texts
-        .iter()
-        .any(|text| text.contains("peers               0/0 connected")));
-    assert!(logs_texts
-        .iter()
-        .any(|text| text.contains("latest warning")));
-    assert!(logs_texts
-        .iter()
-        .any(|text| text.contains("live audit replication marker")));
-    assert!(logs_texts
-        .iter()
-        .any(|text| text.contains("live audit warning marker")));
-    driver.click_target(audit::targets::logs_filter(LogsFilter::Category(
-        crate::telemetry::DesktopLogCategory::Warnings,
-    )));
-    let warning_texts = driver.render();
-    assert!(warning_texts
-        .iter()
-        .any(|text| text.contains("live audit warning marker")));
-    driver.click_target(audit::targets::logs_filter(LogsFilter::Category(
-        crate::telemetry::DesktopLogCategory::Replication,
-    )));
-    let replication_texts = driver.render();
-    assert!(replication_texts
-        .iter()
-        .any(|text| text.contains("live audit replication marker")));
-    assert!(global_log_store().snapshot().total_events > baseline_events);
-
     let returned_chat_texts = driver.open_activity(Activity::Chat);
     assert_eq!(
         driver.app.state.chat.shell.selected_session_id.as_deref(),
@@ -675,7 +377,5 @@ fn desktop_app_live_inference_smoke() -> Result<()> {
 
     runtime.block_on(running_agent.shutdown())?;
     driver.app.shutdown_client();
-    shutdown_core(runtime.as_ref(), peer)?;
     Ok(())
 }
-
