+++ crates/defra-agent-desktop/src/app/tests/live/common.rs
@@ -443,4 +443,3 @@ pub(crate) fn assert_live_deployment_default_config(
         },
     )
 }
-
