+++ crates/defra-agent-desktop/src/app/tests/live/operator_roundtrip.rs
@@ -390,254 +390,3 @@ fn desktop_app_live_operator_config_round_trips() -> Result<()> {
 
     fixture.shutdown()
 }
-
-#[test]
-#[ignore = "hits live inference backend configured by DEFRA_AGENT_DESKTOP_LIVE_BACKEND_* or OPENROUTER_API_KEY"]
-fn desktop_app_live_operator_identity_field_round_trips() -> Result<()> {
-    let _live_guard = live_desktop_test_guard();
-    let mut fixture = build_live_desktop_fixture("audit-live-identities", global_log_store())?;
-    let docs = fixture.docs.clone();
-    let backend = fixture.backend.clone();
-    let agent_did = fixture
-        .running_agent
-        .as_ref()
-        .ok_or_else(|| anyhow!("live fixture missing running agent"))?
-        .did
-        .clone();
-    let shadow_agent_did = format!("{agent_did}:identity-shadow");
-    let behavior_id = format!("{}:identity-behavior", docs.behavior_id);
-    let renamed_behavior_id = format!("{}:renamed", behavior_id);
-    let tool_selection_id = format!("{}:identity-tools", docs.behavior_id);
-    let renamed_tool_selection_id = format!("{}:renamed", tool_selection_id);
-    let profile_id = format!("{}:identity-profile", docs.behavior_id);
-    let renamed_profile_id = format!("{}:renamed", profile_id);
-    let task_id = format!("{}:identity-task", docs.behavior_id);
-    let renamed_task_id = format!("{}:renamed", task_id);
-
-    {
-        let client = Arc::clone(
-            fixture
-                .driver
-                .app
-                .client
-                .as_ref()
-                .ok_or_else(|| anyhow!("desktop client missing"))?,
-        );
-        fixture.runtime.block_on(async {
-            insert_agent_principal(
-                client.as_ref(),
-                &shadow_agent_did,
-                "Identity Shadow",
-                &renamed_behavior_id,
-            )
-            .await?;
-            client
-                .save_tool_selection(&ToolSelectionRow {
-                    selection_id: tool_selection_id.clone(),
-                    agent_did: Some(agent_did.clone()),
-                    display_name: Some("Live Identity Tools".to_string()),
-                    enable_file_tools: Some(false),
-                    file_tools_mode: Some("readonly".to_string()),
-                    enable_bash: Some(false),
-                    bash_mode: Some("disabled".to_string()),
-                    cli_tool_names: vec!["rg".to_string()],
-                    enable_meta_tools: Some(false),
-                    delegate_to: vec![],
-                })
-                .await?;
-            client
-                .save_inference_profile(&InferenceProfileRow {
-                    profile_id: profile_id.clone(),
-                    display_name: Some("Live Identity Profile".to_string()),
-                    context_window: Some(32768),
-                    max_output_tokens: Some(512),
-                    max_turns: Some(8),
-                    temperature: Some(0.0),
-                    stream_batch_ms: Some(40),
-                    deadline_duration_secs: Some(120),
-                })
-                .await?;
-            client
-                .save_behavior(&AgentBehaviorRow {
-                    behavior_id: behavior_id.clone(),
-                    agent_did: Some(agent_did.clone()),
-                    display_name: Some("Live Identity Behavior".to_string()),
-                    system_prompt: Some("Identity field audit behavior.".to_string()),
-                    backend_id: Some(docs.backend_id.clone()),
-                    model_name: Some(backend.model_name.clone()),
-                    tool_selection_id: Some(tool_selection_id.clone()),
-                    inference_profile_id: Some(profile_id.clone()),
-                    compaction_strategy: Some("none".to_string()),
-                    compaction_threshold: Some(0.95),
-                    enabled: Some(false),
-                    created_at: Some(chrono::Utc::now().to_rfc3339()),
-                })
-                .await?;
-            client
-                .save_scheduled_task(&ScheduledTaskRow {
-                    task_id: task_id.clone(),
-                    agent_did: Some(agent_did.clone()),
-                    behavior_id: Some(behavior_id.clone()),
-                    name: Some("Live Identity Task".to_string()),
-                    prompt: Some("Audit identity fields.".to_string()),
-                    interval_secs: Some(3600),
-                    enabled: Some(false),
-                    next_run_at: Some("2035-01-01T00:00:00Z".to_string()),
-                    last_run_at: None,
-                    last_status: None,
-                    last_error: None,
-                    run_count: Some(0),
-                    created_at: Some(chrono::Utc::now().to_rfc3339()),
-                    updated_at: None,
-                })
-                .await?;
-            Ok::<(), anyhow::Error>(())
-        })?;
-    }
-
-    {
-        let driver = &mut fixture.driver;
-        driver.app.state.operator.selected_agent_did = Some(agent_did.clone());
-        driver.app.state.operator.selected_peer_id = None;
-        driver.app.state.operator.selected_entity_id = None;
-        driver.app.state.operator.draft = None;
-        driver.app.state.operator.draft_source_entity_id = None;
-        driver.open_activity(Activity::Operator);
-        driver.click_target(&audit::targets::operator_section(
-            crate::state::OperatorSection::Behaviors,
-        ));
-        driver.wait_for_target(
-            "live identity behavior row",
-            Duration::from_secs(10),
-            &audit::targets::operator_entity(&behavior_id),
-        )?;
-        driver.click_target(&audit::targets::operator_entity(&behavior_id));
-        driver.replace_text_in_target(
-            &audit::targets::operator_field("Behavior ID"),
-            &renamed_behavior_id,
-        );
-        driver.replace_text_in_target(
-            &audit::targets::operator_field("Agent DID"),
-            &shadow_agent_did,
-        );
-        driver.click_target(audit::targets::OPERATOR_APPLY);
-        wait_for_value(
-            "live identity behavior id persisted",
-            Duration::from_secs(5),
-            || {
-                driver.app.client.as_ref().and_then(|client| {
-                    client
-                        .store()
-                        .snapshot()
-                        .behaviors
-                        .iter()
-                        .find(|row| row.behavior_id == renamed_behavior_id)
-                        .filter(|row| row.agent_did.as_deref() == Some(shadow_agent_did.as_str()))
-                        .map(|row| row.behavior_id.clone())
-                })
-            },
-        )?;
-
-        driver.click_target(&audit::targets::operator_section(
-            crate::state::OperatorSection::ToolSelections,
-        ));
-        driver.wait_for_target(
-            "live identity tool selection row",
-            Duration::from_secs(10),
-            &audit::targets::operator_entity(&tool_selection_id),
-        )?;
-        driver.click_target(&audit::targets::operator_entity(&tool_selection_id));
-        driver.replace_text_in_target(
-            &audit::targets::operator_field("Selection ID"),
-            &renamed_tool_selection_id,
-        );
-        driver.click_target(audit::targets::OPERATOR_APPLY);
-        wait_for_value(
-            "live identity tool selection id persisted",
-            Duration::from_secs(5),
-            || {
-                driver.app.client.as_ref().and_then(|client| {
-                    client
-                        .store()
-                        .snapshot()
-                        .tool_selections
-                        .iter()
-                        .find(|row| row.selection_id == renamed_tool_selection_id)
-                        .filter(|row| row.agent_did.as_deref() == Some(agent_did.as_str()))
-                        .map(|row| row.selection_id.clone())
-                })
-            },
-        )?;
-
-        driver.click_target(&audit::targets::operator_section(
-            crate::state::OperatorSection::InferenceProfiles,
-        ));
-        driver.wait_for_target(
-            "live identity profile row",
-            Duration::from_secs(10),
-            &audit::targets::operator_entity(&profile_id),
-        )?;
-        driver.click_target(&audit::targets::operator_entity(&profile_id));
-        driver.replace_text_in_target(
-            &audit::targets::operator_field("Profile ID"),
-            &renamed_profile_id,
-        );
-        driver.click_target(audit::targets::OPERATOR_APPLY);
-        wait_for_value(
-            "live identity profile id persisted",
-            Duration::from_secs(5),
-            || {
-                driver.app.client.as_ref().and_then(|client| {
-                    client
-                        .store()
-                        .snapshot()
-                        .inference_profiles
-                        .iter()
-                        .find(|row| row.profile_id == renamed_profile_id)
-                        .map(|row| row.profile_id.clone())
-                })
-            },
-        )?;
-
-        driver.click_target(&audit::targets::operator_section(
-            crate::state::OperatorSection::ScheduledTasks,
-        ));
-        driver.wait_for_target(
-            "live identity scheduled task row",
-            Duration::from_secs(10),
-            &audit::targets::operator_entity(&task_id),
-        )?;
-        driver.click_target(&audit::targets::operator_entity(&task_id));
-        driver.replace_text_in_target(&audit::targets::operator_field("Task ID"), &renamed_task_id);
-        driver.replace_text_in_target(
-            &audit::targets::operator_field("Behavior ID"),
-            &renamed_behavior_id,
-        );
-        driver.scroll_right_rail_until_target(
-            "live identity scheduled task apply",
-            audit::targets::OPERATOR_APPLY,
-        )?;
-        driver.click_interactable_target(audit::targets::OPERATOR_APPLY)?;
-        wait_for_value(
-            "live identity scheduled task id persisted",
-            Duration::from_secs(5),
-            || {
-                driver.app.client.as_ref().and_then(|client| {
-                    client
-                        .store()
-                        .snapshot()
-                        .scheduled_tasks
-                        .iter()
-                        .find(|row| row.task_id == renamed_task_id)
-                        .filter(|row| {
-                            row.behavior_id.as_deref() == Some(renamed_behavior_id.as_str())
-                        })
-                        .map(|row| row.task_id.clone())
-                })
-            },
-        )?;
-    }
-
-    fixture.shutdown()
-}
-
